+++ src/app/features/styleguide/buttons/buttons.component.html
@@ -1,0 +1,13 @@
+<div class="btn-group aa--mb-2 aa--grid aa--gap-5">
+  <button
+    class="aa--py-2 aa--px-5 aa--me-2 aa--mb-2 aa--text-sm aa--font-bold aa--text-blue-600 aa--bg-white aa--border aa--border-gray-300"
+  >
+    Sign In
+  </button>
+  <button
+    type="button"
+    class="aa--py-2 aa--px-5 aa--me-2 aa--mb-2 aa--text-sm aa--font-bold aa--text-white aa--bg-blue-600"
+  >
+    Post A Job
+  </button>
+</div>
