+++ src/app/features/styleguide/colors/colors.component.html
@@ -1,0 +1,9 @@
+<div class="color-palette aa--leading-9">
+  <div
+    *ngFor="let color of colors"
+    [style.background]="color"
+    class="color-box aa--p-2 aa--m-2 aa--min-w-36 aa--text-center aa--text-lg"
+  >
+    {{ color }}
+  </div>
+</div>
