+++ src/app/features/styleguide/colors/colors.component.ts
@@ -1,11 +1,12 @@
+import { NgFor } from '@angular/common';
 import { Component } from '@angular/core';
 
 @Component({
   selector: 'aa-colors',
-  imports: [],
+  imports: [NgFor],
   templateUrl: './colors.component.html',
   styleUrl: './colors.component.scss'
 })
 export class ColorsComponent {
-
+  colors = ['#2A879B', '#F8FAFD', '#8ADFC4'];
 }
