+++ src/app/features/styleguide/company-cards/company-cards.component.html
@@ -1,0 +1,3 @@
+<div class="card-group aa--grid aa--gap-3">
+  <aa-company-card *ngFor="let card of cards" [companyData]="card" />
+</div>
