+++ src/app/features/styleguide/icons/icons.component.html
@@ -1,0 +1,5 @@
+<div class="icon-group aa--grid aa--gap-5 aa--items-center">
+  <fa-icon [icon]="faMagnifyingGlass" size="1x" class="aa--text-blue-600 aa--px-2" />
+  <fa-icon [icon]="faFacebook" size="2x" class="aa--bg-transparent aa--px-2" />
+  <fa-icon [icon]="faInstagram" size="2x" class="aa--bg-transparent aa--px-2" />
+</div>
