+++ src/app/features/styleguide/icons/icons.component.ts
@@ -1,0 +1,16 @@
+import { Component } from '@angular/core';
+import { FontAwesomeModule } from '@fortawesome/angular-fontawesome';
+import { faFacebook, faInstagram } from '@fortawesome/free-brands-svg-icons';
+import { faMagnifyingGlass } from '@fortawesome/free-solid-svg-icons';
+
+@Component({
+  selector: 'aa-icons',
+  imports: [FontAwesomeModule],
+  templateUrl: './icons.component.html',
+  styleUrl: './icons.component.scss'
+})
+export class IconsComponent {
+  faMagnifyingGlass = faMagnifyingGlass;
+  faFacebook = faFacebook;
+  faInstagram = faInstagram;
+}
