+++ src/app/features/styleguide/job-cards/job-cards.component.html
@@ -1,0 +1,3 @@
+<div class="card-group aa--grid aa--gap-3">
+  <aa-job-card *ngFor="let card of cards" [jobData]="card" />
+</div>
