+++ src/app/features/styleguide/styleguide.component.html
@@ -1,110 +1,63 @@
 <h1 class="aa--font-bold aa--text-lg aa--mb-2">Various Project Styles</h1>
-
-<!-- Colors Section -->
-<h2
-  (click)="toggleSection('colors')"
-  (keydown.enter)="toggleSection('colors')"
-  (keydown.space)="toggleSection('colors')"
-  class="aa--cursor-pointer aa--font-bold aa--underline"
-  tabindex="0"
->
-  Colors
-</h2>
-<ng-template [ngIf]="sectionsVisibility['colors']">
-  <div class="color-palette aa--mb-2 aa--max-w-96">
-    <div *ngFor="let color of colors" [style.background]="color" class="color-box">
-      {{ color }}
-    </div>
-  </div>
-</ng-template>
-
-<!-- Text Sizes Section -->
-<h2
-  (click)="toggleSection('textSizes')"
-  (keydown.enter)="toggleSection('textSizes')"
-  (keydown.space)="toggleSection('textSizes')"
-  class="aa--cursor-pointer aa--font-bold aa--underline"
-  tabindex="0"
->
-  Text Sizes
-</h2>
-<ng-template [ngIf]="sectionsVisibility['textSizes']">
-  <p *ngFor="let size of textSizes" [ngClass]="size.class">
-    Example {{ size.size }} text
-  </p>
-</ng-template>
-
-<!-- Button Variants -->
-<h2
-  (click)="toggleSection('buttons')"
-  (keydown.enter)="toggleSection('buttons')"
-  (keydown.space)="toggleSection('buttons')"
-  class="aa--cursor-pointer aa--font-bold aa--underline"
-  tabindex="0"
->
-  Buttons
-</h2>
-<ng-template [ngIf]="sectionsVisibility['buttons']">
-  <div class="btn-group aa--mb-2">
-    <button
-      class="aa--py-2 aa--px-5 aa--me-2 aa--mb-2 aa--text-sm aa--font-bold aa--text-blue-600 aa--bg-white aa--border aa--border-gray-300"
+<div class="aa--flex aa--gap-2 aa--py-4 aa--px-2 aa--justify-start aa--bg-gray-100">
+  <div class="aa--w-1/3 aa--border-e-2 aa--leading-10 aa--text-wrap">
+    <!-- Colors Section -->
+    <h2
+      class="aa--cursor-pointer aa--font-bold aa--underline"
+      [routerLink]="['colors']"
+      routerLinkActive="aa--text-blue-600"
     >
-      Sign In
-    </button>
-    <button
-      type="button"
-      class="aa--py-2 aa--px-5 aa--me-2 aa--mb-2 aa--text-sm aa--font-bold aa--text-white aa--bg-blue-600"
+      Colors
+    </h2>
+
+    <!-- Text Sizes Section -->
+    <h2
+      class="aa--cursor-pointer aa--font-bold aa--underline"
+      [routerLink]="['textsizes']"
+      routerLinkActive="aa--text-blue-600"
     >
-      Post A Job
-    </button>
-  </div>
-</ng-template>
+      Text Sizes
+    </h2>
 
-<!-- Card Variants -->
-<h2
-  (click)="toggleSection('jobCards')"
-  (keydown.enter)="toggleSection('jobCards')"
-  (keydown.space)="toggleSection('jobCards')"
-  class="aa--cursor-pointer aa--font-bold aa--underline"
-  tabindex="0"
->
-  Job Cards
-</h2>
-<ng-template [ngIf]="sectionsVisibility['jobCards']">
-  <div class="card-group aa--grid aa--gap-3">
-    <aa-job-card *ngFor="let card of cards" [jobData]="card" />
-  </div>
-</ng-template>
+    <!-- Button Variants -->
+    <h2
+      class="aa--cursor-pointer aa--font-bold aa--underline"
+      [routerLink]="['buttons']"
+      routerLinkActive="aa--text-blue-600"
+    >
+      Buttons
+    </h2>
 
-<h2
-  (click)="toggleSection('companyCards')"
-  (keydown.enter)="toggleSection('companyCards')"
-  (keydown.space)="toggleSection('companyCards')"
-  class="aa--cursor-pointer aa--font-bold aa--underline"
-  tabindex="0"
->
-  Company Cards
-</h2>
-<ng-template [ngIf]="sectionsVisibility['companyCards']">
-  <div class="card-group aa--grid aa--gap-3">
-    <aa-company-card *ngFor="let card of cards2" [companyData]="card" />
-  </div>
-</ng-template>
+    <!-- Card Variants -->
+    <!-- 1. Job Cards  -->
+    <h2
+      class="aa--cursor-pointer aa--font-bold aa--underline"
+      [routerLink]="['jobcards']"
+      routerLinkActive="aa--text-blue-600"
+    >
+      Job Cards
+    </h2>
 
-<!-- Icons -->
-<h2
-  (click)="toggleSection('icons')"
-  (keydown.enter)="toggleSection('icons')"
-  (keydown.space)="toggleSection('icons')"
-  class="aa--cursor-pointer aa--font-bold aa--underline"
-  tabindex="0"
->
-  Icons
-</h2>
-<ng-template [ngIf]="sectionsVisibility['icons']">
-  <div class="icon-group">
-    <fa-icon [icon]="faMagnifyingGlass" size="1x" class="aa--text-blue-600 aa--px-2" />
-    <fa-icon [icon]="faFacebook" size="2x" class="aa--bg-transparent aa--px-2" />
-    <fa-icon [icon]="faInstagram" size="2x" class="aa--bg-transparent aa--px-2" />
+    <!-- 2. Company Cards  -->
+    <h2
+      class="aa--cursor-pointer aa--font-bold aa--underline"
+      [routerLink]="['companycards']"
+      routerLinkActive="aa--text-blue-600"
+    >
+      Company Cards
+    </h2>
+
+    <!-- Icons -->
+    <h2
+      class="aa--cursor-pointer aa--font-bold aa--underline"
+      [routerLink]="['icons']"
+      routerLinkActive="aa--text-blue-600"
+    >
+      Icons
+    </h2>
   </div>
-</ng-template>
+
+  <div class="aa--flex aa--items-center aa--justify-center aa--w-2/3">
+    <router-outlet />
+  </div>
+</div>
