+++ src/app/features/styleguide/textsizes/textsizes.component.html
@@ -1,0 +1,3 @@
+<p class="aa--leading-10" *ngFor="let size of textSizes" [ngClass]="size.class">
+  Example {{ size.size }} text
+</p>
