--- src/app/styleguide/buttons/buttons.component.html
@@ -1,1 +1,0 @@
-<p>buttons works!</p>
