--- src/app/styleguide/colors/colors.component.html
@@ -1,1 +1,0 @@
-<p>colors works!</p>
