--- src/app/styleguide/company-cards/company-cards.component.html
@@ -1,1 +1,0 @@
-<p>company-cards works!</p>
