--- src/app/styleguide/company-cards/company-cards.component.ts
@@ -1,11 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'aa-company-cards',
-  imports: [],
-  templateUrl: './company-cards.component.html',
-  styleUrl: './company-cards.component.scss'
-})
-export class CompanyCardsComponent {
-
-}
