--- src/app/styleguide/icons/icons.component.html
@@ -1,1 +1,0 @@
-<p>icons works!</p>
