--- src/app/styleguide/icons/icons.component.ts
@@ -1,11 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'aa-icons',
-  imports: [],
-  templateUrl: './icons.component.html',
-  styleUrl: './icons.component.scss'
-})
-export class IconsComponent {
-
-}
