--- src/app/styleguide/job-cards/job-cards.component.html
@@ -1,1 +1,0 @@
-<p>job-cards works!</p>
