--- src/app/styleguide/job-cards/job-cards.component.ts
@@ -1,11 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'aa-job-cards',
-  imports: [],
-  templateUrl: './job-cards.component.html',
-  styleUrl: './job-cards.component.scss'
-})
-export class JobCardsComponent {
-
-}
