--- src/app/styleguide/textsizes/textsizes.component.html
@@ -1,1 +1,0 @@
-<p>textsizes works!</p>
