--- src/app/styleguide/textsizes/textsizes.component.ts
@@ -1,11 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'aa-textsizes',
-  imports: [],
-  templateUrl: './textsizes.component.html',
-  styleUrl: './textsizes.component.scss'
-})
-export class TextsizesComponent {
-
-}
